backend/requirements.txt
fastapi
uvicorn
mangum
pypdf
pyttsx3
requests
python-dotenv
mistralai
